* PlayerDataService: Fix null error from listFiles (check isDirectory first) * startEvent: Store new event object to field

+++ src/main/java/com/github/ms5984/survivelist/survivelistevents/model/PlayerDataService.java
@@ -56,7 +56,9 @@ public class PlayerDataService {
      * Always called onDisable.
      */
     public static void clearCache(JavaPlugin javaPlugin) {
-        for (File userFile : new File(javaPlugin.getDataFolder(), "users").listFiles()) {
+        final File usersFolder = new File(javaPlugin.getDataFolder(), "users");
+        if (!usersFolder.isDirectory()) return;
+        for (File userFile : usersFolder.listFiles()) {
             if (userFile.isFile()) {
                 //noinspection ResultOfMethodCallIgnored
                 userFile.delete();
